Replace examples with parsed versions
